requirements.txt
# Core Framework
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
pgvector>=0.2.3

# Data Analysis
pandas>=2.1.4
numpy>=1.26.2
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# API & Async
aiohttp>=3.9.1
requests>=2.31.0
websockets>=12.0

# Task Scheduling
apscheduler>=3.10.4
celery>=5.3.4
redis>=5.0.1

# Email
sendgrid>=6.11.0

# Web Framework
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.3

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
